+++ src/routes/(app)/_components/Navbar.svelte
@@ -20,7 +20,8 @@
 					{ label: 'Categories', href: '/admin/categories' },
 					{ label: 'Events', href: '/admin/events' },
 					{ label: 'Submissions', href: '/admin/submissions' },
-					{ label: 'Users', href: '/admin/users' }
+					{ label: 'Users', href: '/admin/users' },
+					{ label: 'Tickets', href: '/admin/tickets' }
 				]
 			: [
 					{ label: 'Home', href: '/' },
@@ -58,9 +59,9 @@
 			{/if}
 		</Button>
 	</div>
-	<HorizontalMenu highlightActive={false}>
+	<HorizontalMenu highlightActive={false} class="max-md:hidden">
 		{#each menuItems as item}
-			<HorizontalMenuItem href={item.href}>{item.label}</HorizontalMenuItem>
+			<HorizontalMenuItem href={item.href}>{$t(item.label)}</HorizontalMenuItem>
 		{/each}
 	</HorizontalMenu>
 
@@ -111,7 +112,7 @@
 
 <!-- Mobile menu -->
 {#if isMenuOpen}
-	<div class="absolute left-0 top-0 z-10 h-full w-full bg-black" transition:fly={{ y: -100 }}>
+	<div class="absolute left-0 top-0 z-10 h-full w-full bg-bg" transition:fly={{ y: -100 }}>
 		<div class="flex h-full items-center justify-center">
 			<ul class="flex flex-col gap-lg text-2xl">
 				{#each menuItems as item}