+++ src/routes/admin/tickets/+page.server.ts
@@ -1,0 +1,13 @@
+import { db } from '$lib/server/db';
+import type { PageServerLoad } from './$types';
+
+export const load = (async () => {
+	const tickets = await db.query.tickets.findMany({
+		with: {
+			user: { columns: { username: true, email: true, id: true } },
+			event: { columns: { name: true, id: true } }
+		}
+	});
+
+	return { tickets, title: { text: 'Tickets', href: '/admin/tickets' } };
+}) satisfies PageServerLoad;