+++ src/routes/admin/users/+page.server.ts
@@ -2,7 +2,9 @@ import { db } from '$lib/server/db';
 import type { PageServerLoad } from './$types';
 
 export const load = (async () => {
-	const users = await db.query.users.findMany();
+	const users = await db.query.users.findMany({
+		with: { tickets: { columns: { id: true } }, submissions: { columns: { id: true } } }
+	});
 
-	return { users };
+	return { users, title: { text: 'Users', href: '/admin/users' } };
 }) satisfies PageServerLoad;